backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
regex>=2023.10.3
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
alembic>=1.12.0
openai>=1.0.0
pillow>=10.0.0
httpx>=0.25.0
aiofiles>=23.0.0